% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Significant_Features.R
\name{find_concordant_features_across_sites}
\alias{find_concordant_features_across_sites}
\title{Generate target vector containing all significant features in all DC vs all comparisons}
\usage{
find_concordant_features_across_sites(filepath_to_significant_results_tsv)
}
\arguments{
\item{filepath_to_significant_results_tsv}{a filepath to significant_results.tsv output from Maaslin2}
}
\value{
a character vector containing all concordant and significant features across six sites
}
\description{
Generate target vector containing all significant features in all DC vs all comparisons
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
