% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa_Site_Heatmap.R
\name{find_features_union_for_type_heatmap}
\alias{find_features_union_for_type_heatmap}
\title{Generating heatmaps for "GBM" - gut brain modules}
\usage{
find_features_union_for_type_heatmap(
  duo_filepath,
  jej_filepath,
  ile_filepath,
  cec_filepath,
  pc_filepath,
  dc_filepath
)
}
\arguments{
\item{path_to_all_results_tsv}{filepath to Maaslin2/all_results.tsv}

\item{targetvector}{vector containing the featureIDs that you want to include in the heatmap. these featureIDs must also be in all_results.tsv}

\item{path_to_module_key}{filepath to the GMM key with annotations for each GMM}

\item{titlestring}{title of plot, e.g. "Mucosal"}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating heatmaps for "GBM" - gut brain modules
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
