% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_Type_Heatmap.R
\name{generate_GMM_heat_map_by_Type}
\alias{generate_GMM_heat_map_by_Type}
\title{Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules}
\usage{
generate_GMM_heat_map_by_Type(
  dataframe,
  path_to_Module_Key,
  Y,
  ystring,
  titlestring,
  colorvector,
  breakvector
)
}
\arguments{
\item{dataframe}{dataframe containing Maaslin2 results across all six LMEM}

\item{Y}{hierachy_annotations, feature_annotations, metabolic_map, or Map}

\item{ystring}{choose between "hierachy_annotations", "feature_annotations", "metabolic_map", or "Map"}

\item{titlestring}{title of plot, e.g. "Mucosal"}

\item{colorvector}{vector of colors}

\item{breakvector}{vector of breaks}

\item{path_to_module_key}{filepath to the GMM key with annotations for each GMM}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
