% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_Interregional_Barplot.R
\name{generate_interregional_GBM_barplot}
\alias{generate_interregional_GBM_barplot}
\title{Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules}
\usage{
generate_interregional_GBM_barplot(
  path_to_all_results_tsv,
  path_to_Module_Key,
  titlestring,
  colorvector
)
}
\arguments{
\item{path_to_all_results_tsv}{filepath to Maaslin2/all_results.tsv}

\item{titlestring}{title of plot, e.g. "Mucosal"}

\item{colorvector}{character vector of two colors}

\item{path_to_module_key}{filepath to the GMM key with annotations for each GMM}

\item{ystring}{choose between "hierachy_annotations", "feature_annotations", "metabolic_map", or "Map"}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
