% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_Interregional_Barplot.R
\name{generate_interregional_GMM_barplot_shotgun}
\alias{generate_interregional_GMM_barplot_shotgun}
\title{Generating interregional GMM barplot specific to shotgun data}
\usage{
generate_interregional_GMM_barplot_shotgun(
  path_to_all_results_tsv,
  path_to_Module_Key,
  ystring,
  titlestring,
  colorvector
)
}
\arguments{
\item{path_to_all_results_tsv}{filepath to Maaslin2/all_results.tsv}

\item{path_to_Module_Key}{filepath to the GMM key with annotations for each GMM}

\item{ystring}{choose between "hierachy_annotations", "feature_annotations", "metabolic_map", or "Map"}

\item{titlestring}{title of plot, e.g. "Mucosal"}

\item{colorvector}{character vector of two colors}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating interregional GMM barplot specific to shotgun data
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
