% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa_Interregional_by_Type.R
\name{generate_interregional_taxa_barplot_shotgun}
\alias{generate_interregional_taxa_barplot_shotgun}
\title{Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules}
\usage{
generate_interregional_taxa_barplot_shotgun(
  path_to_significant_results_tsv,
  titlestring,
  colorvector
)
}
\arguments{
\item{path_to_significant_results_tsv}{filepath to Maaslin2/significant_results.tsv}

\item{titlestring}{title of plot, e.g. "Mucosal"}

\item{colorvector}{character vector of two colors}

\item{ystring}{choose between "hierachy_annotations", "feature_annotations", "metabolic_map", or "Map"}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating heatmaps to match the metabolic maps containing "GMM" - gut metabolic modules
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
