% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa_Barplots.R
\name{generate_L6_taxa_plots}
\alias{generate_L6_taxa_plots}
\title{Generate taxa summary plots by genera showing only the top 0.1\% most abundant}
\usage{
generate_L6_taxa_plots(
  path_to_RDS,
  titlestring,
  greppattern,
  fillvector,
  graphby
)
}
\arguments{
\item{path_to_RDS}{filepath to where your L6 RDS is stored}

\item{titlestring}{title of plot}

\item{greppattern}{".*s__" for species}

\item{fillvector}{a named vector of colors}

\item{graphby}{a string, can pass "Site" or "Type"}
}
\value{
a ggplot2 object which you can then further customize by adding ggplot2 functions
}
\description{
Plot does not have legend - legends were graphed separately
}
\examples{

generate_L6_taxa_plots("CS-Facility-Analysis/Taxa-Barplots/Luminal_level-6.RDS", 
"Luminal ( > 0.1\% Relative Abundance)",
 ".*g__",
  assign_cols,
   "Site")


}
\author{
Julianne Yang
}
