% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_Diversity.R
\name{generate_pcoA_plots}
\alias{generate_pcoA_plots}
\title{Generate PCoA plots for beta diversity}
\usage{
generate_pcoA_plots(
  ordination_file,
  metadata,
  title,
  colorvariable,
  colorvector
)
}
\arguments{
\item{title}{string containing title of plot}

\item{colorvariable}{string containing the variable you want to color the points by. Available options are "Site", "Type", or "Site_General".}

\item{colorvector}{named character vector with colors assigned to levels of colorvariable}

\item{input_data}{ordination file directly from QIIME2 principal_coordinates.py. read in with header=FALSE}

\item{input_metadata}{dataframe containing "SampleID" exactly as written}
}
\value{
a ggplot2 object encoding a pcoA plot
}
\description{
(Site_General, Site, Type)
}
\author{
Julianne C. Yang

Jonathan P. Jacobs
}
