% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa_Site_Heatmap.R
\name{generate_taxa_heat_map_by_site}
\alias{generate_taxa_heat_map_by_site}
\title{Generating heatmaps for "GBM" - gut brain modules}
\usage{
generate_taxa_heat_map_by_site(
  path_to_all_results_tsv,
  targetvector,
  titlestring,
  colorvector,
  breakvector
)
}
\arguments{
\item{path_to_all_results_tsv}{filepath to Maaslin2/all_results.tsv}

\item{targetvector}{vector containing the featureIDs that you want to include in the heatmap. these featureIDs must also be in all_results.tsv}

\item{titlestring}{title of plot, e.g. "Mucosal"}

\item{path_to_module_key}{filepath to the GMM key with annotations for each GMM}
}
\value{
a ggplot2 object encoding a heatmap
}
\description{
Generating heatmaps for "GBM" - gut brain modules
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
