% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa_Barplots.R
\name{get_phyla_from_plot}
\alias{get_phyla_from_plot}
\title{Generate taxa summary plots by phyla
Plot does not have legend - legends were graphed separately}
\usage{
get_phyla_from_plot(filepath)
}
\arguments{
\item{filepath}{filepath to the csv file with wrangled names}
}
\value{
unique character vector of genera which will go in the plot
}
\description{
Generate taxa summary plots by phyla
Plot does not have legend - legends were graphed separately
}
\examples{

generate_L2_taxa_plots("CS-Facility-Analysis/Taxa-Barplots/Luminal_level-6.RDS", 
"Luminal ( > 0.1\% Relative Abundance)",
 ".*g__",
  assign_cols,
   "Site")


}
\author{
Julianne Yang
}
