% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coef_Line_Plots.R
\name{process_results_files}
\alias{process_results_files}
\title{Data wrangling for the coefficient Line Plots}
\usage{
process_results_files(
  file_paths,
  feature_value,
  new_value = "Distal_Colon",
  new_coef = 0,
  cohort_prefixes,
  filter_by = "Site"
)
}
\arguments{
\item{file_paths}{filepaths to dataframes containing statistical testing results}

\item{feature_value}{string descriptor of feature name}

\item{new_value}{string descriptor of the reference variable}

\item{new_coef}{integer value for the effect size of the reference variable}

\item{cohort_prefixes}{strings describing what's in each dataframe}
}
\value{
a dataframe aggregating all results dataframes with cohort prefixes appended
}
\description{
Data wrangling for the coefficient Line Plots
}
\author{
Julianne C. Yang

Jonathan P. Jacobs
}
