% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upset_Plots.R
\name{process_results_for_upset_plot}
\alias{process_results_for_upset_plot}
\title{Generate Upset Plots for GBM and GMM}
\usage{
process_results_for_upset_plot(file_paths, cohort_prefixes)
}
\arguments{
\item{file_paths}{filepaths to dataframes containing statistical testing results}

\item{cohort_prefixes}{strings describing what's in each dataframe}
}
\value{
a dataframe aggregating all results dataframes with cohort prefixes appended
}
\description{
Generate Upset Plots for GBM and GMM
}
\author{
Julianne C. Yang

Jonathan P. Jacobs
}
