% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Significant_Features.R
\name{query_type_features_union}
\alias{query_type_features_union}
\title{Query a vector of features against all entries within multiple dataframes}
\usage{
query_type_features_union(
  target,
  duo_filepath,
  jej_filepath,
  ile_filepath,
  cec_filepath,
  pc_filepath,
  dc_filepath
)
}
\arguments{
\item{target}{vector generated from find_features_union_for_type_heatmap}
}
\value{
a character vector containing all concordant and significant features across six sites
}
\description{
Query a vector of features against all entries within multiple dataframes
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
