% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Repeated_Measures_PERMANOVA.R
\name{run_repeated_PERMANOVA}
\alias{run_repeated_PERMANOVA}
\title{Use PERMANOVA with repeat measure-aware permutations from Lloyd Price et al 2019.}
\usage{
run_repeated_PERMANOVA(
  path_to_distance_matrix_tsv,
  path_to_metadata_csv,
  permute_columns_vector,
  subject_metadata_vector
)
}
\arguments{
\item{path_to_distance_matrix_tsv}{filepath to distance_matrix.tsv}

\item{path_to_metadata_csv}{filepath to metadata.csv}

\item{permute_columns_vector}{a character vector containing names of metadata columns that repeat per subject (e.g. Timepoint)}

\item{subject_metadata_vector}{a character vector containing names of metadata columns consistent for each subject with MouseID as the last element (e.g. Age, Sex, MouseID)}
}
\value{
res.aov object from adonis
}
\description{
Use PERMANOVA with repeat measure-aware permutations from Lloyd Price et al 2019.
}
\author{
Julianne C. Yang, Jonathan P. Jacobs
}
